services/clickstream/backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
pydantic-settings==2.5.2
boto3==1.35.31
pyiceberg[s3fs]==0.20.0
ray[default]==2.37.0
pandas==2.2.3
httpx==0.27.2
python-multipart==0.0.9
numpy==1.26.4

